+++ LotteryTicketWebAPI/RequestProcessing/SQLRequest.cs
@@ -5,9 +5,9 @@ internal readonly struct SQLRequest
     internal static string GetAllTicketsQuery()
         => $"SELECT * FROM dbo.AllTickets;";
 
-    internal static string GetTicketAnIdQuery(int? id)
+    internal static string GetTicketAnIdQuery(long? id)
         => $"SELECT * FROM dbo.AllTickets WHERE Id = {id};";
 
-    internal static string BuyTicketAnIdQuery(int? id)
+    internal static string BuyTicketAnIdQuery(long? id)
         => $"SELECT * FROM dbo.AllTickets WHERE Id = {id};";
 }
